Ajustes para autenticación y registro de usuarios.
Descripción de ajustes:

1. Edición de línea de importación de la base de datos en src\app\api\users\route.js

Antes:

import connectToDatabase from "../../../libs/db_user";

Ahora:

import connectToDatabase from "../../../libs/db_users";

2. Edición de línea de importación de la base de datos en src\app\api\users\[id]\route.js

Antes:

import connectToDatabase from "../../../../libs/db_user";

Ahora:

import connectToDatabase from "../../../../libs/db_users";

3. Edición de archivo src\components\UserForm.jsx se ajusta código para registro de usuarios y login. En este archivo se adicionan las líneas:

import countryOptions from '../data/countryOptions';   // Importar countryOptions
import phoneLengthsByCountry from '../data/phoneLengthsByCountry';   // Importar phoneLengthsByCountry

4. Se crean los archivos src\data\countryOptions.jsx y src\data\phoneLengthsByCountry.jsx, donde se ubican respectivamente las validaciones de código de país y de número de teléfono. Tales validaciones antes estaban en el archivo src\components\UserForm.jsx.

5. Edición de src\pages\api\auth\[...nextauth].js.

Además de Google y Facebook, se agrega un proveedor de Credentials para autenticar a los usuarios mediante su correo electrónico y contraseña.

Antes:

providers: [
  GoogleProvider({ ... }),
  FacebookProvider({ ... }),
]

Ahora:

Ahora:

providers: [
  CredentialsProvider({ ... }), // Añadido
  GoogleProvider({ ... }),
  FacebookProvider({ ... }),
]

Se usa el proveedor CredentialsProvider con una función authorize que autentica a los usuarios a través de la base de datos, verificando que el correo y la contraseña sean correctos.

// authenticate con credenciales
async authorize(credentials) {
  const client = await clientPromise;
  const db = client.db();
  const user = await db.collection("users").findOne({ email: credentials.email });

  if (user && await bcrypt.compare(credentials.password, user.password)) {
    return {
      id: user._id.toString(),
      name: `${user.firstname} ${user.secondname}`,
      email: user.email,
      role: user.role,
    };
  }
  return null;
}

Se tenía sólo el callback redirect para redirigir al usuario después de iniciar sesión. Se adicionan 2 callbacks:

session: Añade información adicional al objeto session, como el role del usuario, que no estaba presente en la primera configuración.

jwt: Almacena los datos del usuario (id, email, role) en el token JWT.

Se importa bcryptjs para realizar la comparación de la contraseña de los usuarios en el proveedor de CredentialsProvider.

import bcrypt from "bcryptjs";

Propósito de la Configuración: Añade autenticación mediante credenciales (correo y contraseña) almacenadas en la base de datos, junto con una mayor personalización en los callbacks para manejar datos de sesión y tokens JWT.

La configuración ajustada es más completa, ya que soporta tanto autenticación con proveedores externos como con credenciales personalizadas.

6. Se agrega la dependencia dotenv@16.4.5.

dotenv es un paquete de Node.js utilizado para cargar variables de entorno desde un archivo .env a process.env. Esto es útil para gestionar configuraciones sensibles, como claves API, credenciales de bases de datos, y otras variables que no deberían estar hardcodeadas en el código fuente.

 On branch enrique
 Changes to be committed:
	modified:   package-lock.json
	modified:   package.json
	modified:   src/app/api/users/[id]/route.js
	modified:   src/app/api/users/route.js
	modified:   src/components/UserForm.jsx
	new file:   src/data/countryOptions.jsx
	new file:   src/data/phoneLengthsByCountry.jsx
	modified:   src/pages/api/auth/[...nextauth].js
	new file:   src/pages/api/users/create-user.js

 Changes not staged for commit:
	modified:   .env
