+++ package.json
@@ -9,8 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@radix-ui/react-dialog": "^1.1.2", // Nueva dependencia Henry
     "@next-auth/mongodb-adapter": "^1.1.3",
+    "@radix-ui/react-dialog": "^1.1.2",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-radio-group": "^1.2.1",
@@ -18,29 +18,29 @@
     "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.1",
-    "@radix-ui/react-toast": "^1.2.2",
+    "@radix-ui/themes": "^3.1.5",
     "axios": "^1.7.7",
+    "bcryptjs": "^2.4.3",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "dotenv": "^16.4.5",
+    "dotenv": "16.4.5",
     "libphonenumber-js": "^1.11.12",
     "lucide-react": "^0.452.0",
-    "mongodb": "^6.10.0", // Actualiza dependencia
-    "next-auth": "^4.24.10", // Nueva dependencia Enrique
-    "mongoose": "^8.8.0", // Nueva dependencia Elkin
-    "next": "^15.0.2", // Actualiza dependencia
-    "nextjs": "^0.0.3", // Nueva dependencia Elkin
+    "mongodb": "5.9.2",
+    "mongoose": "7.8.2",
+    "next": "14.2.15",
+    "next-auth": "4.24.10",
     "react": "^18",
     "react-country-flag": "^3.1.0",
     "react-dom": "^18",
     "react-icons": "^5.3.0",
-    "tailwind-merge": "^2.5.3",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwind-merge": "2.5.3",
+    "tailwindcss-animate": "1.0.7"
   },
   "devDependencies": {
     "eslint": "^8",
     "eslint-config-next": "14.2.15",
     "postcss": "^8",
-    "tailwindcss": "^3.4.1"
+    "tailwindcss": "3.4.1"
   }
 }
