+++ src/data/countryOptions.jsx
@@ -1,0 +1,156 @@
+// Validación de código de país
+const countryOptions = [
+  { code: "+93", country: "Afganistán", flag: "AF", region: "Asia" },
+  { code: "+355", country: "Albania", flag: "AL", region: "Europa" },
+  { code: "+49", country: "Alemania", flag: "DE", region: "Europa" },
+  { code: "+213", country: "Argelia", flag: "DZ", region: "África" },
+  { code: "+376", country: "Andorra", flag: "AD", region: "Europa" },
+  { code: "+244", country: "Angola", flag: "AO", region: "África" },
+  { code: "+1264", country: "Anguila", flag: "AI", region: "Caribe" },
+  { code: "+1268", country: "Antigua y Barbuda", flag: "AG", region: "Caribe" },
+  { code: "+54", country: "Argentina", flag: "AR", region: "Sudamérica" },
+  { code: "+374", country: "Armenia", flag: "AM", region: "Asia" },
+  { code: "+297", country: "Aruba", flag: "AW", region: "Caribe" },
+  { code: "+61", country: "Australia", flag: "AU", region: "Oceanía" },
+  { code: "+43", country: "Austria", flag: "AT", region: "Europa" },
+  { code: "+994", country: "Azerbaiyán", flag: "AZ", region: "Asia" },
+  { code: "+1242", country: "Bahamas", flag: "BS", region: "Caribe" },
+  { code: "+973", country: "Baréin", flag: "BH", region: "Asia" },
+  { code: "+880", country: "Bangladés", flag: "BD", region: "Asia" },
+  { code: "+1246", country: "Barbados", flag: "BB", region: "Caribe" },
+  { code: "+32", country: "Bélgica", flag: "BE", region: "Europa" },
+  { code: "+501", country: "Belice", flag: "BZ", region: "América Central" },
+  { code: "+229", country: "Benín", flag: "BJ", region: "África" },
+  { code: "+1441", country: "Bermudas", flag: "BM", region: "Caribe" },
+  { code: "+975", country: "Bután", flag: "BT", region: "Asia" },
+  { code: "+591", country: "Bolivia", flag: "BO", region: "Sudamérica" },
+  { code: "+387", country: "Bosnia y Herzegovina", flag: "BA", region: "Europa" },
+  { code: "+267", country: "Botsuana", flag: "BW", region: "África" },
+  { code: "+55", country: "Brasil", flag: "BR", region: "Sudamérica" },
+  { code: "+673", country: "Brunéi", flag: "BN", region: "Asia" },
+  { code: "+359", country: "Bulgaria", flag: "BG", region: "Europa" },
+  { code: "+226", country: "Burkina Faso", flag: "BF", region: "África" },
+  { code: "+257", country: "Burundi", flag: "BI", region: "África" },
+  { code: "+238", country: "Cabo Verde", flag: "CV", region: "África" },
+  { code: "+855", country: "Camboya", flag: "KH", region: "Asia" },
+  { code: "+237", country: "Camerún", flag: "CM", region: "África" },
+  { code: "+1", country: "Canadá", flag: "CA", region: "América del Norte" },
+  { code: "+235", country: "Chad", flag: "TD", region: "África" },
+  { code: "+56", country: "Chile", flag: "CL", region: "Sudamérica" },
+  { code: "+86", country: "China", flag: "CN", region: "Asia" },
+  { code: "+357", country: "Chipre", flag: "CY", region: "Europa" },
+  { code: "+57", country: "Colombia", flag: "CO", region: "Sudamérica" },
+  { code: "+269", country: "Comoras", flag: "KM", region: "África" },
+  { code: "+242", country: "Congo", flag: "CG", region: "África" },
+  { code: "+243", country: "Congo (RDC)", flag: "CD", region: "África" },
+  { code: "+850", country: "Corea del Norte", flag: "KP", region: "Asia" },
+  { code: "+82", country: "Corea del Sur", flag: "KR", region: "Asia" },
+  { code: "+506", country: "Costa Rica", flag: "CR", region: "América Central" },
+  { code: "+225", country: "Costa de Marfil", flag: "CI", region: "África" },
+  { code: "+385", country: "Croacia", flag: "HR", region: "Europa" },
+  { code: "+53", country: "Cuba", flag: "CU", region: "Caribe" },
+  { code: "+45", country: "Dinamarca", flag: "DK", region: "Europa" },
+  { code: "+1767", country: "Dominica", flag: "DM", region: "Caribe" },
+  { code: "+1809", country: "República Dominicana", flag: "DO", region: "Caribe" },
+  { code: "+593", country: "Ecuador", flag: "EC", region: "Sudamérica" },
+  { code: "+20", country: "Egipto", flag: "EG", region: "África" },
+  { code: "+503", country: "El Salvador", flag: "SV", region: "América Central" },
+  { code: "+971", country: "Emiratos Árabes Unidos", flag: "AE", region: "Asia" },
+  { code: "+291", country: "Eritrea", flag: "ER", region: "África" },
+  { code: "+421", country: "Eslovaquia", flag: "SK", region: "Europa" },
+  { code: "+386", country: "Eslovenia", flag: "SI", region: "Europa" },
+  { code: "+34", country: "España", flag: "ES", region: "Europa" },
+  { code: "+372", country: "Estonia", flag: "EE", region: "Europa" },
+  { code: "+251", country: "Etiopía", flag: "ET", region: "África" },
+  { code: "+63", country: "Filipinas", flag: "PH", region: "Asia" },
+  { code: "+358", country: "Finlandia", flag: "FI", region: "Europa" },
+  { code: "+33", country: "Francia", flag: "FR", region: "Europa" },
+  { code: "+241", country: "Gabón", flag: "GA", region: "África" },
+  { code: "+220", country: "Gambia", flag: "GM", region: "África" },
+  { code: "+995", country: "Georgia", flag: "GE", region: "Asia" },
+  { code: "+233", country: "Ghana", flag: "GH", region: "África" },
+  { code: "+350", country: "Gibraltar", flag: "GI", region: "Europa" },
+  { code: "+30", country: "Grecia", flag: "GR", region: "Europa" },
+  { code: "+1473", country: "Granada", flag: "GD", region: "Caribe" },
+  { code: "+1671", country: "Guam", flag: "GU", region: "Oceanía" },
+  { code: "+502", country: "Guatemala", flag: "GT", region: "América Central" },
+  { code: "+245", country: "Guinea-Bisáu", flag: "GW", region: "África" },
+  { code: "+592", country: "Guyana", flag: "GY", region: "Sudamérica" },
+  { code: "+509", country: "Haití", flag: "HT", region: "Caribe" },
+  { code: "+504", country: "Honduras", flag: "HN", region: "América Central" },
+  { code: "+852", country: "Hong Kong", flag: "HK", region: "Asia" },
+  { code: "+36", country: "Hungría", flag: "HU", region: "Europa" },
+  { code: "+91", country: "India", flag: "IN", region: "Asia" },
+  { code: "+62", country: "Indonesia", flag: "ID", region: "Asia" },
+  { code: "+98", country: "Irán", flag: "IR", region: "Asia" },
+  { code: "+964", country: "Irak", flag: "IQ", region: "Asia" },
+  { code: "+353", country: "Irlanda", flag: "IE", region: "Europa" },
+  { code: "+44", country: "Reino Unido", flag: "GB", region: "Europa" },
+  { code: "+972", country: "Israel", flag: "IL", region: "Asia" },
+  { code: "+39", country: "Italia", flag: "IT", region: "Europa" },
+  { code: "+1876", country: "Jamaica", flag: "JM", region: "Caribe" },
+  { code: "+81", country: "Japón", flag: "JP", region: "Asia" },
+  { code: "+962", country: "Jordania", flag: "JO", region: "Asia" },
+  { code: "+7", country: "Kazajistán", flag: "KZ", region: "Asia" },
+  { code: "+254", country: "Kenia", flag: "KE", region: "África" },
+  { code: "+686", country: "Kiribati", flag: "KI", region: "Oceanía" },
+  { code: "+965", country: "Kuwait", flag: "KW", region: "Asia" },
+  { code: "+996", country: "Kirguistán", flag: "KG", region: "Asia" },
+  { code: "+856", country: "Laos", flag: "LA", region: "Asia" },
+  { code: "+371", country: "Letonia", flag: "LV", region: "Europa" },
+  { code: "+961", country: "Líbano", flag: "LB", region: "Asia" },
+  { code: "+266", country: "Lesoto", flag: "LS", region: "África" },
+  { code: "+231", country: "Liberia", flag: "LR", region: "África" },
+  { code: "+218", country: "Libia", flag: "LY", region: "África" },
+  { code: "+423", country: "Liechtenstein", flag: "LI", region: "Europa" },
+  { code: "+370", country: "Lituania", flag: "LT", region: "Europa" },
+  { code: "+352", country: "Luxemburgo", flag: "LU", region: "Europa" },
+  { code: "+853", country: "Macau", flag: "MO", region: "Asia" },
+  { code: "+261", country: "Madagascar", flag: "MG", region: "África" },
+  { code: "+265", country: "Malaui", flag: "MW", region: "África" },
+  { code: "+60", country: "Malasia", flag: "MY", region: "Asia" },
+  { code: "+960", country: "Maldivas", flag: "MV", region: "Asia" },
+  { code: "+223", country: "Malí", flag: "ML", region: "África" },
+  { code: "+356", country: "Malta", flag: "MT", region: "Europa" },
+  { code: "+692", country: "Islas Marshall", flag: "MH", region: "Oceanía" },
+  { code: "+596", country: "Martinica", flag: "MQ", region: "Caribe" },
+  { code: "+222", country: "Mauritania", flag: "MR", region: "África" },
+  { code: "+230", country: "Mauricio", flag: "MU", region: "África" },
+  { code: "+52", country: "México", flag: "MX", region: "América del Norte" },
+  { code: "+691", country: "Micronesia", flag: "FM", region: "Oceanía" },
+  { code: "+373", country: "Moldavia", flag: "MD", region: "Europa" },
+  { code: "+377", country: "Mónaco", flag: "MC", region: "Europa" },
+  { code: "+976", country: "Mongolia", flag: "MN", region: "Asia" },
+  { code: "+382", country: "Montenegro", flag: "ME", region: "Europa" },
+  { code: "+1", country: "Montserrat", flag: "MS", region: "Caribe" },
+  { code: "+212", country: "Marruecos", flag: "MA", region: "África" },
+  { code: "+258", country: "Mozambique", flag: "MZ", region: "África" },
+  { code: "+264", country: "Namibia", flag: "NA", region: "África" },
+  { code: "+674", country: "Nauru", flag: "NR", region: "Oceanía" },
+  { code: "+977", country: "Nepal", flag: "NP", region: "Asia" },
+  { code: "+31", country: "Países Bajos", flag: "NL", region: "Europa" },
+  { code: "+599", country: "Curazao", flag: "CW", region: "Caribe" },
+  { code: "+64", country: "Nueva Zelanda", flag: "NZ", region: "Oceanía" },
+  { code: "+505", country: "Nicaragua", flag: "NI", region: "América Central" },
+  { code: "+227", country: "Níger", flag: "NE", region: "África" },
+  { code: "+234", country: "Nigeria", flag: "NG", region: "África" },
+  { code: "+683", country: "Niue", flag: "NU", region: "Oceanía" },
+  { code: "+850", country: "Corea del Norte", flag: "KP", region: "Asia" },
+  { code: "+47", country: "Noruega", flag: "NO", region: "Europa" },
+  { code: "+968", country: "Omán", flag: "OM", region: "Asia" },
+  { code: "+92", country: "Pakistán", flag: "PK", region: "Asia" },
+  { code: "+680", country: "Palaos", flag: "PW", region: "Oceanía" },
+  { code: "+970", country: "Palestina", flag: "PS", region: "Asia" },
+  { code: "+507", country: "Panamá", flag: "PA", region: "América Central" },
+  { code: "+675", country: "Papúa Nueva Guinea", flag: "PG", region: "Oceanía" },
+  { code: "+595", country: "Paraguay", flag: "PY", region: "Sudamérica" },
+  { code: "+51", country: "Perú", flag: "PE", region: "Sudamérica" },
+  { code: "+63", country: "Filipinas", flag: "PH", region: "Asia" },
+  { code: "+48", country: "Polonia", flag: "PL", region: "Europa" },
+  { code: "+1", country: "Puerto Rico", flag: "PR", region: "Caribe" },
+  { code: "+1", country: "República Dominicana", flag: "DO", region: "Caribe" },
+  { code: "+1", country: "San Cristóbal y Nieves", flag: "KN", region: "Caribe" },
+  { code: "+1", country: "San Vicente y las Granadinas", flag: "VC", region: "Caribe" },
+];
+
+export default countryOptions;
